New tool: CLI Command Editor (#108)
What's CLI command editor?
In my day-to-day work, I have to use a lot of CLI commands. I often copy and paste them to my terminal and edited the command option values as required. Observing that many people face the similar issue, which can get tad bit annoying for complex commands, I have made this tool which easily lets you modify the option values.

How it works?
Paste the command.
A form will be presented with all the options and flags, you can edit this form.
Copy the command that gets generated in real-time as you edit it.